requirements.txt
streamlit==1.31.0
torch==2.6.0
transformers==4.36.0
pymongo==4.6.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.3
sentencepiece==0.1.99
protobuf==3.20.3
accelerate==0.24.1
pdf2image==1.16.3
python-docx==1.0.0
PyPDF2==3.0.1
plotly==5.18.0
bcrypt==4.0.1
python-jose==3.3.0
requests==2.31.0